+++ mne/channels/tests/test_montage.py
@@ -560,4 +560,40 @@ def _check_roundtrip(montage, fname):
     assert_equal(montage_read.coord_frame, 'head')
 
 
+from mne.utils import object_diff
+from mne.io import read_raw_nicolet
+from mne.channels import Montage
+
+nicolet_fname = op.join(io_dir, 'nicolet', 'tests', 'data',
+                        'test_nicolet_raw.data')
+
+
+def _fake_montage(ch_names):
+    return Montage(
+        pos=np.random.RandomState(42).randn(len(ch_names), 3),
+        ch_names=ch_names,
+        kind='foo',
+        selection=np.arange(len(ch_names))
+    )
+
+
+@pytest.mark.parametrize('read_raw,fname', [
+    pytest.param(read_raw_nicolet, nicolet_fname, id='nicolet'),
+])
+def test_montage_when_reading_and_setting(read_raw, fname):
+    """Test montage."""
+    raw_none = read_raw(input_fname=fname, ch_type='eeg',
+                        montage=None, preload=False)
+    montage = _fake_montage(raw_none.info['ch_names'])
+
+    raw_montage = read_raw(input_fname=fname, ch_type='eeg',
+                           montage=montage, preload=False)
+    raw_none.set_montage(montage)
+
+    # Check they are the same
+    assert_array_equal(raw_none.get_data(), raw_montage.get_data())
+    assert object_diff(raw_none.info['dig'], raw_montage.info['dig']) == ''
+    assert object_diff(raw_none.info['chs'], raw_montage.info['chs']) == ''
+
+
 run_tests_if_main()
